Turn menu creation methods into extension methods
We're going to need to create menus from more places.
Also consolidate extensions some more, and remove
unused menu command id
