+++ src/GitHub.Exports/Services/Services.cs
@@ -100,15 +100,5 @@ namespace GitHub.VisualStudio
                 return null;
             return GitService.GitServiceHelper.GetUri(solutionDir);
         }
-
-        public static IRepository GetRepoFromSolution(this IVsSolution solution)
-        {
-            string solutionDir, solutionFile, userFile;
-            if (!ErrorHandler.Succeeded(solution.GetSolutionInfo(out solutionDir, out solutionFile, out userFile)))
-                return null;
-            if (solutionDir == null)
-                return null;
-            return GitService.GitServiceHelper.GetRepo(solutionDir);
-        }
     }
 }
